@@ -11,48 +11,54 @@
 # limitations under the License.
 # ===----------------------------------------------------------------------=== #
 
-import os
+# DOC: max/tutorials/build-custom-ops.mdx
+
 from pathlib import Path
 
 import numpy as np
 from max.driver import CPU, Accelerator, Tensor, accelerator_count
 from max.dtype import DType
 from max.engine import InferenceSession
-from max.graph import Graph, TensorType, ops
+from max.graph import DeviceRef, Graph, TensorType, ops
 
 if __name__ == "__main__":
-    # This is necessary only in specific build environments.
-    if directory := os.getenv("BUILD_WORKSPACE_DIRECTORY"):
-        os.chdir(directory)
-
-    path = Path(__file__).parent / "kernels.mojopkg"
+    mojo_kernels = Path(__file__).parent / "kernels"
 
     rows = 5
     columns = 10
     dtype = DType.float32
-
+    # Place the graph on a GPU, if available. Fall back to CPU if not.
+    device = CPU() if accelerator_count() == 0 else Accelerator()
     # Configure our simple one-operation graph.
     graph = Graph(
         "addition",
         # The custom Mojo operation is referenced by its string name, and we
         # need to provide inputs as a list as well as expected output types.
         forward=lambda x: ops.custom(
-            name="add_one_custom",
+            name="add_one",
+            device=DeviceRef.from_device(device),
             values=[x],
-            out_types=[TensorType(dtype=x.dtype, shape=x.tensor.shape)],
+            out_types=[
+                TensorType(
+                    dtype=x.dtype,
+                    shape=x.tensor.shape,
+                    device=DeviceRef.from_device(device),
+                )
+            ],
         )[0].tensor,
         input_types=[
-            TensorType(dtype, shape=[rows, columns]),
+            TensorType(
+                dtype,
+                shape=[rows, columns],
+                device=DeviceRef.from_device(device),
+            ),
         ],
+        custom_extensions=[mojo_kernels],
     )
-
-    # Place the graph on a GPU, if available. Fall back to CPU if not.
-    device = CPU() if accelerator_count() == 0 else Accelerator()
 
     # Set up an inference session for running the graph.
     session = InferenceSession(
         devices=[device],
-        custom_extensions=path,
     )
 
     # Compile the graph.
@@ -61,10 +67,10 @@ if __name__ == "__main__":
     # Fill an input matrix with random values.
     x_values = np.random.uniform(size=(rows, columns)).astype(np.float32)
 
-    # Create a driver tensor from this, and move it to the accelerator.
+    # Create a tensor and move it to the device (CPU or GPU).
     x = Tensor.from_numpy(x_values).to(device)
 
-    # Perform the calculation on the target device.
+    # Run inference with the input tensor.
     result = model.execute(x)[0]
 
     # Copy values back to the CPU to be read.
