@@ -11,25 +11,23 @@
 # limitations under the License.
 # ===----------------------------------------------------------------------=== #
 
-import os
 from pathlib import Path
 
 import numpy as np
 from max.driver import CPU, Accelerator, Tensor, accelerator_count
 from max.dtype import DType
 from max.engine import InferenceSession
-from max.graph import Graph, TensorType, ops
+from max.graph import DeviceRef, Graph, TensorType, ops
 
 if __name__ == "__main__":
-    # This is necessary only in specific build environments.
-    if directory := os.getenv("BUILD_WORKSPACE_DIRECTORY"):
-        os.chdir(directory)
-
-    path = Path(__file__).parent / "kernels.mojopkg"
+    mojo_kernels = Path(__file__).parent / "kernels"
 
     rows = 5
     columns = 10
     dtype = DType.float32
+
+    # Place the graph on a GPU, if available. Fall back to CPU if not.
+    device = CPU() if accelerator_count() == 0 else Accelerator()
 
     # Configure our simple one-operation graph.
     graph = Graph(
@@ -39,23 +37,31 @@ if __name__ == "__main__":
         # Since the custom operation is parametric, we need to provide the
         # parameters as a dictionary.
         forward=lambda x: ops.custom(
-            name="add_constant_custom",
+            name="add_constant",
+            device=DeviceRef.from_device(device),
             values=[x],
-            out_types=[TensorType(dtype=x.dtype, shape=x.tensor.shape)],
+            out_types=[
+                TensorType(
+                    dtype=x.dtype,
+                    shape=x.tensor.shape,
+                    device=DeviceRef.from_device(device),
+                )
+            ],
             parameters={"value": 5},
         )[0].tensor,
         input_types=[
-            TensorType(dtype, shape=[rows, columns]),
+            TensorType(
+                dtype,
+                shape=[rows, columns],
+                device=DeviceRef.from_device(device),
+            ),
         ],
+        custom_extensions=[mojo_kernels],
     )
-
-    # Place the graph on a GPU, if available. Fall back to CPU if not.
-    device = CPU() if accelerator_count() == 0 else Accelerator()
 
     # Set up an inference session for running the graph.
     session = InferenceSession(
         devices=[device],
-        custom_extensions=path,
     )
 
     # Compile the graph.
